# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "This is the API documentation for Safar API",
        "contact": {
            "email": "zahra77babaee@gmail.com.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/ticket/{id}": {
            "get": {
                "tags": [
                    "Tickets"
                ],
                "summary": "نمایش تیکت خاص",
                "description": "دریافت اطلاعات یک تیکت و پیام‌های مرتبط با آن",
                "operationId": "d282d288209a69efe0ce6a93dde6b90e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "شناسه تیکت مورد نظر",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "توکن احراز هویت JWT",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "Bearer {your_token}"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "اطلاعات تیکت و پیام‌ها با موفقیت بازگردانده شد",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "OK"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "تیکت با موفقیت پیدا شد."
                                        },
                                        "data": {
                                            "properties": {
                                                "ticket_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "title": {
                                                    "type": "string",
                                                    "example": "مشکل در ورود به حساب"
                                                },
                                                "priority": {
                                                    "type": "string",
                                                    "example": "high"
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "open"
                                                },
                                                "creator_id": {
                                                    "type": "integer",
                                                    "example": 2
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2024-10-25 14:30"
                                                },
                                                "messages": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "message_id": {
                                                                "type": "integer",
                                                                "example": 10
                                                            },
                                                            "creator_id": {
                                                                "type": "integer",
                                                                "example": 2
                                                            },
                                                            "message": {
                                                                "type": "string",
                                                                "example": "پیام کاربر یا ادمین"
                                                            },
                                                            "is_read": {
                                                                "type": "boolean",
                                                                "example": false
                                                            },
                                                            "created_at": {
                                                                "type": "string",
                                                                "format": "date-time",
                                                                "example": "2024-10-25 15:30"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "کاربر مجاز به مشاهده این تیکت نیست",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "ERROR"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "شما دسترسی به این تیکت را ندارید."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "تیکت مورد نظر یافت نشد",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "ERROR"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "تیکت مورد نظر یافت نشد."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/tickets": {
            "get": {
                "tags": [
                    "Tickets"
                ],
                "summary": "دریافت لیست تیکت‌ها",
                "description": "این متد لیست تیکت‌ها را بر اساس نقش کاربر (مدیر یا کاربر) نمایش می‌دهد.",
                "operationId": "687cb3d4766d4523828bf9265c460503",
                "responses": {
                    "200": {
                        "description": "لیست تیکت‌ها با موفقیت دریافت شد",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "ok"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "title": {
                                                        "type": "string",
                                                        "example": "مشکل در ورود به سایت"
                                                    },
                                                    "priority": {
                                                        "type": "string",
                                                        "example": "high"
                                                    },
                                                    "status": {
                                                        "type": "string",
                                                        "example": "open"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
...
}